refactor: move sidebar endpoint and handler to plugin router and controller respectively
