+++ backend/controllers/plugin.controller.js
@@ -1,13 +1,15 @@
 const axios = require('axios');
-const { name, description, version, author, license, bugs: issues, homepage } = require('../../package.json');
+const { name, version, author, license, bugs: issues } = require('../../package.json');
+const pluginId = name;
+const pluginName = 'Company Files Management Plug-In';
 
 exports.info = async (req, res) => {
   const repoInfo = await axios.get('https://api.github.com/repos/zurichat/zc_plugin_company_files');
 
   res.status(200).json({
-    pluginId: name,
-    pluginName: description.split('~').pop().trim(),
-    description,
+    pluginId,
+    pluginName,
+    description: 'An effective file management system that improves business workflow, organizes important data and provides a searchable database for quick retrieval.',
     version,
     author,
     repo: {
@@ -25,4 +27,40 @@ exports.info = async (req, res) => {
       lastUpdated: repoInfo.data.updated_at
     },
   });
+}
+
+exports.sidebar = (req,res) => {
+  const sidebarListObject = {
+      pluginId,
+      pluginName,
+      organisationId: 'jd93-ene82-rt92-df',
+      menuLink:{
+        title: 'Company Files',
+        icon: '<URL to image>',
+        action: 'Open Plugin'
+        },
+      subItems:[{
+        title: 'All Files',
+        icon: '<URL to image>',
+        action: 'View All Files'
+      },{
+        title: 'Shared',
+        icon: '<URL to image>',
+        action: 'View All Files'
+      },{
+        title: 'Starred Files',
+        icon: '<URL to image>',
+        action: 'View Starred Files'
+      },{
+        title: 'Trash',
+        icon: '<URL to image>',
+        action: 'View Trash'
+      },{
+        title: 'Help',
+        icon: '<URL to image>',
+        action: 'View Help'
+      }]
+  };
+
+  res.status(200).json(sidebarListObject);
 }
