+++ backend/routes/plugin.router.js
@@ -1,6 +1,7 @@
 const router = require('express').Router();
-const { info } = require('../controllers/plugin.controller');
+const { info, sidebar } = require('../controllers/plugin.controller');
 
 router.get('/info', info);
+router.get('/sidebar', sidebar);
 
 module.exports = router;
